+++ NEEngine/Data/Shaders/VertexSphericalDopplerShader.hlsl
@@ -27,7 +27,7 @@ struct VertexShaderOutput
 {
 	float2 tex : TEXCOORD0;
 	float fogFactor : FOG_FACTOR;
-	double velocity : VELOCITY;
+	float velocity : VELOCITY;
 	float4 position : SV_POSITION; //должно быть последним при поступлении в пиксельный шейдер, если в нем не будем его брать (иначе всё сместится)
 };
